[PyTorch Mobile] Generate selected_kernel_dtypes.h during the build
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49279

Now that the YAML files for tracing based selective build optionally have thee information regarding the selected kernel function dtypes, we can start generating an include file (`selected_kernel_dtypes.h`) to make the inclusion of code for specific dtypes selective based on compile time decisions.

The way this is done is that if we detect that the code for a specific dtype should not be in the binary, we add an exception (throw) statement just before the method is called (see the first diff in this dtack) and allow the compiler to optimize away the rest of the function's body. This has the advantage of allowing the compiler to know the lambda's return type (since it's inferred from the `return` statements in the body of the method, and if we compile out all the cases, then the compiler won't know the return type and it will result in a compilation error).

The generated `selected_kernel_dtypes.h` file isn't yet used anywhere since we can't make the kernel implementations selective yet (since ATen/native isn't yet selective per-app).

When merging, we need to handle the case of both older and newer (tracing based) operator lists. If we detect any operator that includes all overloads, it indicates that an old style operator list is part of the build, and we need to `include_all_kernel_dtypes` for this build.
ghstack-source-id: 118692416

Differential Revision: [D25388949](https://our.internmc.facebook.com/intern/diff/D25388949/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D25388949/)!