[Try] [PyTorch Mobile] Skip signature check when converting to typed operator handle
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49469

In Functions.cpp, there is a call to `typed<...>` that converts to a `TypedOperatorHandle`. This isn't needed at runtime since it's already been exercised during development, and for mobile, there is no possibility of operators or kernels being registered by users (from Python code the way it is possible on server side).
ghstack-source-id: 118714246

Differential Revision: [D25581159](https://our.internmc.facebook.com/intern/diff/D25581159/)