quant: ensure observers do not crash for empty Tensors
Summary:

Ensures that having a Tensor with 0 elements does not crash observers.
Note: it's illegal to pass Tensors with 0 elements to reductions such
as min and max, so we gate this out before the logic hits min/max.

This should not be hit often in practice, but it's coming up
during debugging of some RCNN models with test inputs.

Test Plan:

```
python test/test_quantization.py TestObserver.test_zero_numel
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 385c3182ee0772f522f333623e4fdc7a91af36b6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49800